update approve job, send email, manage user

+++ database/migrations/2024_11_19_140128_add_user_id_to_job_posts_table.php
@@ -6,11 +6,19 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    public function up(): void
+    // public function up(): void
+    // {
+    //     Schema::table('job_posts', function (Blueprint $table) {
+    //         $table->foreignId('user_id')->constrained('users')->onDelete('cascade')->after('id');
+    //     });
+    // }
+    public function up()
     {
-        Schema::table('job_posts', function (Blueprint $table) {
-            $table->foreignId('user_id')->constrained('users')->onDelete('cascade')->after('id');
-        });
+        if (!Schema::hasColumn('job_posts', 'user_id')) {
+            Schema::table('job_posts', function (Blueprint $table) {
+                $table->foreignId('user_id')->constrained('users')->onDelete('cascade')->after('id');
+            });
+        }
     }
 
     public function down(): void
